EmptyIntervals.java
You are given an XML document containing elements with the tag name <bar> that are time intervals (containing startdate and enddate). These intervals correspond to the times that different people are working.
The task is to find all the empty intervals (during which no people are working). E.g if the file contains two intervals, where one person is working from 9:00 to 14:00, and another person is working from 16:00 to 19:00, then the interval from 14:00 to 16:00 would be empty since no one is working at that time.
Using Java, write a program that will take this XML document as input and find all the empty intervals where no one is working.
Hint: there could be no such empty intervals.